Extract websocket from api to own file
