+++ index.html
@@ -9,7 +9,7 @@
 </head>
 
 <body style="font-family: Roboto"> <!--TODO: Remove to production -->
-  <div class="crowdsorcerer-widget" peer-review data-assignment='1'></div>
+  <div class="crowdsorcerer-widget" data-assignment='1'></div>
     <script>
       document.addEventListener("DOMContentLoaded", function() {
         window.initCrowdsorcerer();
